Extract Viewer.snapshot into src/render/snapshot.ts behind a structural host
Move the snapshot capture pipeline and its four overlay-compositing draw
helpers (inspector card, probe readout, scale bar, colorbar) out of Viewer.ts
into snapshot.ts. Viewer keeps a one-line delegate that binds its own render
state to a SnapshotHost interface, the same seam exportAdapter and colorLegend
use. Behaviour is unchanged: same option handling, same fast path, same Blob
output, same toBlob-null reject message.

Add tests/snapshot.test.ts (13 cases) covering option resolution, the
fast-path decision and the toBlob-null error through a fake host; the
overlay-compositing branch still needs a real canvas and stays on the e2e
suite. Bank the lower Viewer.ts count (6,959 to 6,458) and reconcile the
truth docs.
