+++ docs/validation/monolith-size-baseline.json
@@ -5,7 +5,7 @@
       "goal": 2500
     },
     "src/render/Viewer.ts": {
-      "lines": 6959,
+      "lines": 6458,
       "goal": 2000
     }
   }
